+++ boot/pom.xml
@@ -38,6 +38,11 @@
         </dependency>
         <dependency>
             <groupId>pt.estga</groupId>
+            <artifactId>content-import</artifactId>
+            <version>${stonemark.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>pt.estga</groupId>
             <artifactId>report</artifactId>
             <version>${stonemark.version}</version>
         </dependency>
