+++ content-import/src/main/java/pt/estga/contentimport/config/ContentImportAutoConfiguration.java
@@ -1,9 +1,7 @@
-package pt.estga.contentimport;
+package pt.estga.contentimport.config;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.persistence.autoconfigure.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @AutoConfiguration
 @ComponentScan("pt.estga.contentimport")
