+++ content-import/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-pt.estga.contentimport.ContentImportAutoConfiguration
+pt.estga.contentimport.config.ContentImportAutoConfiguration
